Adding the ability to add emojis in channel messages (not included in direct messages yet).
I have never actually worked with react before so chances are there is most likely a better and simplier way to do this. I assume what's done from the message reactions can probably be reused some way and have it implemented for channels and messages, feedback is much appreciated.

I also tried including the space's emojis too, however I wasn't sure how to get it displaying in the textbox or get it displayed when it is sent as when choosing it from the emoji picker it just turned out to be a code (combination of characters and dashes)